Fixed issue with checkboxes not updating state
